requirements.txt
python-dotenv
langchain
langchain-core
langchain-community
langchain-openai
langchain-groq
langchain-huggingface
langchainhub
langchain_experimental
langchain-anthropic
langchain-text-splitters
chromadb
langchain-chroma
langchain-text-splitters
esprima
tree_sitter 
tree_sitter_languages
pysqlite3-binary
streamlit